Cast std::size_t to int in order to silence warning

+++ src/server/sockets/GameThread.cpp
@@ -101,7 +101,7 @@ userId GameThread::startGame(const ClientInformations& player1, const ClientInfo
 	packet.clear();
 	std::random_device device;
 	std::mt19937 generator{device()};
-	cardId earnedCardId{std::uniform_int_distribution<int>(0, nbSpells + nbCreatures)(generator)};
+	cardId earnedCardId{std::uniform_int_distribution<int>(0, static_cast<int>(nbSpells + nbCreatures))(generator)};
 	++earnedCardId;  // Card indices start to 1 because of SQLite
 	// \TODO: add earnedCardId to winner's card collection
 	packet << TransferType::GAME_OVER << TransferType::WINNER << static_cast<sf::Int32>(earnedCardId);
